Move text extraction to the resources app
This is a trade-off to enforce the new "no circular dependencies" rule.
To keep resources decoupled from content_search, a content field is added
to AbstractResource. The "update content" and "invoke extractor" endpoints
are to be in resources now, with content_search's resource post-save hook
copying the content field and indexing it.

Although this will work, the trade-off is that "content" is now duplicated.
With proper deferral of this field when querying resources, this should not
be an issue, but it is important to keep in mind.
