+++ solution/backend/resources/serializers/__init__.py
@@ -3,3 +3,4 @@ from .citations import *  # noqa
 from .groups import *  # noqa
 from .subjects import *  # noqa
 from .resources import *  # noqa
+from .text_extraction import *  # noqa
