+++ solution/backend/resources/serializers/text_extraction.py
@@ -1,0 +1,5 @@
+from rest_framework import serializers
+
+
+class ContentUpdateSerializer(serializers.Serializer):
+    text = serializers.CharField()
